add test for short and long
